Resaved the project using OpenShot's latest daily build
Reverted the changes made using the custom branch and resaved the
project using the latest daily build that makes all the projects
portable (along with the thumbnails). The thumbnails now display
correctly, and this whole thing is backwards compatible (tested
with v2.4.1). The project is now ready to be shared with the world.